refactor(widget): extract i18n dictionaries, retranslate and UI builders
Split ~2900 lines out of the PolygonExtractionWidget god-class into three new
modules:

- polygon_widget/ui/i18n_content.py: tooltip/help dictionaries
- polygon_widget/ui/retranslate.py: giant _retranslate_ui method body
- polygon_widget/ui/builders.py: _build_ui and per-tab builder methods
- polygon_widget/ui/pipeline_list.py: PipelineListWidget subclass

widget.py shrinks from 7057 to 4214 lines with no behavior changes; the
extracted functions are bound back onto the widget via thin delegators so all
self.* attribute access remains identical.

Also adds tests/unit/test_widget_smoke.py with a golden snapshot of the public
API to guard the ongoing refactor against accidental drift.

+++ contour/polygon_widget/domain/geometry.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
+from collections.abc import Iterable
 from math import ceil, floor, hypot
-from typing import Iterable
 
 from .polygon import Point
 
@@ -12,14 +12,14 @@ def compute_polygon_metrics(points: Iterable[Point]) -> tuple[float, float, tupl
         return 0.0, 0.0, (0, 0, 0, 0)
     if len(vertices) == 1:
         x_coord, y_coord = vertices[0]
-        return 0.0, 0.0, (int(floor(x_coord)), int(floor(y_coord)), 1, 1)
+        return 0.0, 0.0, (floor(x_coord), floor(y_coord), 1, 1)
     if len(vertices) == 2:
         (x0, y0), (x1, y1) = vertices
         perimeter = hypot(x1 - x0, y1 - y0) * 2.0
-        x_min = int(floor(min(x0, x1)))
-        y_min = int(floor(min(y0, y1)))
-        x_max = int(ceil(max(x0, x1)))
-        y_max = int(ceil(max(y0, y1)))
+        x_min = floor(min(x0, x1))
+        y_min = floor(min(y0, y1))
+        x_max = ceil(max(x0, x1))
+        y_max = ceil(max(y0, y1))
         return 0.0, perimeter, (x_min, y_min, max(1, x_max - x_min), max(1, y_max - y_min))
 
     area_sum = 0.0
@@ -31,8 +31,8 @@ def compute_polygon_metrics(points: Iterable[Point]) -> tuple[float, float, tupl
 
     x_values = [point[0] for point in vertices]
     y_values = [point[1] for point in vertices]
-    x_min = int(floor(min(x_values)))
-    y_min = int(floor(min(y_values)))
-    x_max = int(ceil(max(x_values)))
-    y_max = int(ceil(max(y_values)))
+    x_min = floor(min(x_values))
+    y_min = floor(min(y_values))
+    x_max = ceil(max(x_values))
+    y_max = ceil(max(y_values))
     return abs(area_sum) / 2.0, perimeter, (x_min, y_min, max(1, x_max - x_min + 1), max(1, y_max - y_min + 1))
